requirements.txt
pillow
numpy
torchvision
torchaudio
timm
pandas
scikit-learn
seaborn
matplotlib 
thop 
h5py 
SimpleITK 
scikit-image
tqdm
torchattacks
albumentations==1.3.1
efficientnet-pytorch
pydicom
python-gdcm
pyjpegls
pylibjpeg
dicom-csv